refactoring: mini
