+++ product-service/src/database/config.ts
@@ -1,10 +1,8 @@
-import { Client } from "pg";
-
-export const client = new Client({
+export const DB_Config = {
 	host: process.env.PG_HOST,
 	port: process.env.PG_PORT,
 	database: process.env.PG_NAME,
 	user: process.env.PG_USER,
 	password: process.env.PG_PASS,
 	connectionTimeoutMillis: 5000,
-});
+};
