+++ product-service/src/functions/getProductsList/handler.ts
@@ -1,43 +1,42 @@
+import { Client } from "pg";
 import { formatJSONResponse, formatErrorResponse } from "@libs/apiGateway";
 import { middyfy } from "@libs/lambda";
-import { ValidatedEventAPIGatewayProxyEvent } from "@libs/types";
 
-import { client } from "src/database/client";
-import schema from "../schema";
+import { HandlerType } from "src/types";
+import { DB_Config } from "src/database/config";
 
-const getProductsList: ValidatedEventAPIGatewayProxyEvent<typeof schema> =
-	async () => {
-		await client.connect();
+const getProductsList: HandlerType<{}> = async () => {
+	const client = new Client(DB_Config);
+	await client.connect();
 
-		try {
-			const allProducts = await client.query(
-				`SELECT product.id, product.title, product.description, product.price, stock.count
+	try {
+		const { rows: allProducts } = await client.query(
+			`SELECT product.id, product.title, product.description, product.price, stock.count
 				FROM product, stock
 					WHERE product.id = stock.product_id`
-			);
+		);
 
-			if (allProducts) {
-				return formatJSONResponse({
-					data: allProducts.rows,
-				});
-			}
-
-			return formatErrorResponse({
-				errorMessage: "Products not found",
-				statusCode: 404,
+		if (allProducts) {
+			return formatJSONResponse({
+				data: allProducts,
 			});
-		} catch (error) {
-			const errorMessage =
-				"Something went wrong when looking for product";
-			console.error(errorMessage, error);
-
-			return formatErrorResponse({
-				errorMessage,
-				statusCode: 400,
-			});
-		} finally {
-			client.end();
 		}
-	};
+
+		return formatErrorResponse({
+			errorMessage: "Products not found",
+			statusCode: 404,
+		});
+	} catch (error) {
+		const errorMessage = "Something went wrong when get all products";
+		console.error(errorMessage, error);
+
+		return formatErrorResponse({
+			errorMessage,
+			statusCode: 400,
+		});
+	} finally {
+		client.end();
+	}
+};
 
 export const main = middyfy(getProductsList);
