+++ product-service/src/functions/getProductsList/index.ts
@@ -1,5 +1,4 @@
 import { handlerPath } from "@libs/handlerResolver";
-import schema from "../schema";
 
 export default {
 	handler: `${handlerPath(__dirname)}/handler.main`,
@@ -8,13 +7,6 @@ export default {
 			http: {
 				method: "get",
 				path: "products",
-				responses: {
-					"200": {
-						content: {
-							"application/json": schema,
-						},
-					},
-				},
 			},
 		},
 	],
