--- product-service/src/functions/schema.ts
@@ -1,11 +1,0 @@
-export default {
-	type: "object",
-	properties: {
-		id: { type: "string" },
-		title: { type: "string" },
-		description: { type: "string" },
-		price: { type: "number" },
-		count: { type: "number" },
-		image: { type: "string" },
-	},
-} as const;
