+++ product-service/src/types.ts
@@ -3,14 +3,9 @@ import type {
 	APIGatewayProxyResult,
 	Handler,
 } from "aws-lambda";
-import type { FromSchema } from "json-schema-to-ts";
 
-type ValidatedAPIGatewayProxyEvent<S> = Omit<APIGatewayProxyEvent, "body"> & {
-	body: FromSchema<S>;
-};
-
-export type ValidatedEventAPIGatewayProxyEvent<S> = Handler<
-	ValidatedAPIGatewayProxyEvent<S>,
+export type HandlerType<EventParam> = Handler<
+	APIGatewayProxyEvent & EventParam,
 	APIGatewayProxyResult
 >;
